@@ -4,11 +4,12 @@
 //
 // The goal is to make sure that the division does not fail to compile
 // and returns the proper type.
+// Execute `rustlings hint using_as` or use the `hint` watch subcommand for a hint.
 
 // I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
-    let total = values.iter().fold(0.0, |a, b| a + b);
+    let total = values.iter().sum::<f64>();
     total / values.len()
 }
 
