@@ -1,13 +1,11 @@
 // enums1.rs
-// Make me compile! Execute `rustlings hint enums1` for hints!
+// No hints this time! ;)
+
+// I AM NOT DONE
 
 #[derive(Debug)]
 enum Message {
     // TODO: define a few types of messages as used below
-    Quit,
-    Echo,
-    Move,
-    ChangeColor,
 }
 
 fn main() {
