@@ -1,27 +1,16 @@
 // enums2.rs
-// Make me compile! Execute `rustlings hint enums2` for hints!
+// Execute `rustlings hint enums2` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 #[derive(Debug)]
 enum Message {
     // TODO: define the different variants used below
-    Quit,
-    Echo { msg: String },
-    Move { x: i32, y: i32 },
-    ChangeColor {r: i32, g: i32, b: i32 }
 }
 
 impl Message {
     fn call(&self) {
-        println!("{:?}", &self);
-    }
-
-    fn Echo(&self, msg: &str) {
-        // println!("{:?}", msg);
-    }
-
-    fn ChangeColor(&self, r :i32, g :i32, b :i32) {
-        // println!("{:?}", r, g, b);
+        println!("{:?}", self);
     }
 }
 
