@@ -1,12 +1,13 @@
 // functions2.rs
-// Make me compile! Execute `rustlings hint functions2` for hints :)
+// Execute `rustlings hint functions2` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 fn main() {
     call_me(3);
 }
 
-fn call_me(num: i32) {
+fn call_me(num:) {
     for i in 0..num {
         println!("Ring! Call number {}", i + 1);
     }
