@@ -1,8 +1,10 @@
 // functions3.rs
-// Make me compile! Execute `rustlings hint functions3` for hints :)
+// Execute `rustlings hint functions3` or use the `hint` watch subcommand for a hint.
+
+// I AM NOT DONE
 
 fn main() {
-    call_me(5);
+    call_me();
 }
 
 fn call_me(num: u32) {
