@@ -1,7 +1,7 @@
 // This shopping list program isn't compiling!
 // Use your knowledge of generics to fix it.
 
-// Execute `rustlings hint generics1` for hints!
+// Execute `rustlings hint generics1` or use the `hint` watch subcommand for a hint.
 
 // I AM NOT DONE
 
