@@ -1,16 +1,13 @@
 // if1.rs
+// Execute `rustlings hint if1` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 pub fn bigger(a: i32, b: i32) -> i32 {
     // Complete this function to return the bigger number!
     // Do not use:
     // - another function call
     // - additional variables
-    // Execute `rustlings hint if1` for hints
-    if (a > b) {
-        return a
-    }
-    return b
 }
 
 // Don't mind this for now :)
