@@ -1,12 +1,11 @@
 // modules2.rs
 // You can bring module paths into scopes and provide new names for them with the
 // 'use' and 'as' keywords. Fix these 'use' statements to make the code compile.
-// Make me compile! Execute `rustlings hint modules2` for hints :)
+// Execute `rustlings hint modules2` or use the `hint` watch subcommand for a hint.
 
 // I AM NOT DONE
 
 mod delicious_snacks {
-
     // TODO: Fix these use statements
     use self::fruits::PEAR as ???
     use self::veggies::CUCUMBER as ???
