@@ -3,7 +3,7 @@
 // and especially from the Rust standard library into your scope.
 // Bring SystemTime and UNIX_EPOCH
 // from the std::time module. Bonus style points if you can do it with one line!
-// Make me compile! Execute `rustlings hint modules3` for hints :)
+// Execute `rustlings hint modules3` or use the `hint` watch subcommand for a hint.
 
 // I AM NOT DONE
 
