@@ -1,12 +1,13 @@
 // move_semantics2.rs
-// Make me compile without changing line 13!
-// Execute `rustlings hint move_semantics2` for hints :)
+// Make me compile without changing line 13 or moving line 10!
+// Execute `rustlings hint move_semantics2` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 fn main() {
     let vec0 = Vec::new();
 
-    let mut vec1 = fill_vec(&vec0);
+    let mut vec1 = fill_vec(vec0);
 
     // Do not change the following line!
     println!("{} has length {} content `{:?}`", "vec0", vec0.len(), vec0);
@@ -16,8 +17,8 @@ fn main() {
     println!("{} has length {} content `{:?}`", "vec1", vec1.len(), vec1);
 }
 
-fn fill_vec(vec: &Vec<i32>) -> Vec<i32> {
-    let mut vec = vec.clone();
+fn fill_vec(vec: Vec<i32>) -> Vec<i32> {
+    let mut vec = vec;
 
     vec.push(22);
     vec.push(44);
