@@ -1,7 +1,9 @@
 // move_semantics3.rs
 // Make me compile without adding new lines-- just changing existing lines!
 // (no lines with multiple semicolons necessary!)
-// Execute `rustlings hint move_semantics3` for hints :)
+// Execute `rustlings hint move_semantics3` or use the `hint` watch subcommand for a hint.
+
+// I AM NOT DONE
 
 fn main() {
     let vec0 = Vec::new();
@@ -15,7 +17,7 @@ fn main() {
     println!("{} has length {} content `{:?}`", "vec1", vec1.len(), vec1);
 }
 
-fn fill_vec(mut vec: Vec<i32>) -> Vec<i32> {
+fn fill_vec(vec: Vec<i32>) -> Vec<i32> {
     vec.push(22);
     vec.push(44);
     vec.push(66);
