@@ -1,14 +1,15 @@
 // move_semantics5.rs
-// Make me compile only be reordering the lines in `main()`, but without
+// Make me compile only by reordering the lines in `main()`, but without
 // adding, changing or removing any of them.
-// Execute `rustlings hint move_semantics5` for hints :)
+// Execute `rustlings hint move_semantics5` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 fn main() {
     let mut x = 100;
     let y = &mut x;
-    *y += 100;
     let z = &mut x;
+    *y += 100;
     *z += 1000;
     assert_eq!(x, 1200);
 }
