@@ -2,6 +2,8 @@
 // Fill in the rest of the line that has code missing!
 // No hints, there's no tricks, just get used to typing these :)
 
+// I AM NOT DONE
+
 fn main() {
     // Booleans (`bool`)
 
@@ -10,7 +12,7 @@ fn main() {
         println!("Good morning!");
     }
 
-    let is_evening = false; // Finish the rest of this line like the example! Or make it be false!
+    let // Finish the rest of this line like the example! Or make it be false!
     if is_evening {
         println!("Good evening!");
     }
