@@ -2,9 +2,13 @@
 // Fill in the rest of the line that has code missing!
 // No hints, there's no tricks, just get used to typing these :)
 
+// I AM NOT DONE
+
 fn main() {
     // Characters (`char`)
 
+    // Note the _single_ quotes, these are different from the double quotes
+    // you've been seeing around.
     let my_first_initial = 'C';
     if my_first_initial.is_alphabetic() {
         println!("Alphabetical!");
@@ -14,7 +18,7 @@ fn main() {
         println!("Neither alphabetic nor numeric!");
     }
 
-    let your_character = ')'; // Finish this line like the example! What's your favorite character?
+    let // Finish this line like the example! What's your favorite character?
     // Try a letter, try a number, try a special character, try a character
     // from a different language than your own, try an emoji!
     if your_character.is_alphabetic() {
