@@ -1,10 +1,11 @@
 // primitive_types3.rs
 // Create an array with at least 100 elements in it where the ??? is.
-// Execute `rustlings hint primitive_types3` for hints!
+// Execute `rustlings hint primitive_types3` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 fn main() {
-    let a = [10,11,12];
+    let a = ???
 
     if a.len() >= 100 {
         println!("Wow, that's a big array!");
