@@ -1,13 +1,14 @@
 // primitive_types4.rs
 // Get a slice out of Array a where the ??? is so that the test passes.
-// Execute `rustlings hint primitive_types4` for hints!!
+// Execute `rustlings hint primitive_types4` or use the `hint` watch subcommand for a hint.
+
+// I AM NOT DONE
 
 #[test]
 fn slice_out_of_array() {
     let a = [1, 2, 3, 4, 5];
 
-    // Create slice out of the array
-    let nice_slice = &a[1..4];
+    let nice_slice = ???
 
     assert_eq!([2, 3, 4], nice_slice)
 }
