@@ -1,11 +1,12 @@
 // primitive_types5.rs
 // Destructure the `cat` tuple so that the println will work.
-// Execute `rustlings hint primitive_types5` for hints!
+// Execute `rustlings hint primitive_types5` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 fn main() {
     let cat = ("Furry McFurson", 3.5);
-    let (name, age) = cat;
+    let /* your pattern here */ = cat;
 
     println!("{} is {} years old.", name, age);
 }
