@@ -2,29 +2,29 @@
 // This is a quiz for the following sections:
 // - Variables
 // - Functions
+// - If
 
-// Mary is buying apples. One apple usually costs 2 Rustbucks, but if you buy
-// more than 40 at once, each apple only costs 1! Write a function that calculates
-// the price of an order of apples given the quantity bought. No hints this time!
+// Mary is buying apples. The price of an apple is calculated as follows:
+// - An apple costs 2 rustbucks.
+// - If Mary buys more than 40 apples, each apple only costs 1 rustbuck!
+// Write a function that calculates the price of an order of apples given
+// the quantity bought. No hints this time!
 
+// I AM NOT DONE
 
 // Put your function here!
-fn calculate_apple_price(mut num: i32) -> i32 {
-    if num > 40 {
-        return num
-    } else {
-        return num * 2
-    }
-}
+// fn calculate_price_of_apples {
 
 // Don't modify this function!
 #[test]
 fn verify_test() {
-    let price1 = calculate_apple_price(35);
-    let price2 = calculate_apple_price(40);
-    let price3 = calculate_apple_price(65);
+    let price1 = calculate_price_of_apples(35);
+    let price2 = calculate_price_of_apples(40);
+    let price3 = calculate_price_of_apples(41);
+    let price4 = calculate_price_of_apples(65);
 
     assert_eq!(70, price1);
     assert_eq!(80, price2);
-    assert_eq!(65, price3);
+    assert_eq!(41, price3);
+    assert_eq!(65, price4);
 }
