@@ -1,29 +1,62 @@
 // quiz2.rs
 // This is a quiz for the following sections:
 // - Strings
+// - Vecs
+// - Move semantics
+// - Modules
+// - Enums
 
-// Ok, here are a bunch of values-- some are `String`s, some are `&str`s. Your
-// task is to call one of these two functions on each value depending on what
-// you think each value is. That is, add either `string_slice` or `string`
-// before the parentheses on each line. If you're right, it will compile!
+// Let's build a little machine in form of a function.
+// As input, we're going to give a list of strings and commands. These commands
+// determine what action is going to be applied to the string. It can either be:
+// - Uppercase the string
+// - Trim the string
+// - Append "bar" to the string a specified amount of times
+// The exact form of this will be:
+// - The input is going to be a Vector of a 2-length tuple,
+//   the first element is the string, the second one is the command.
+// - The output element is going to be a Vector of strings.
+// No hints this time!
 
+// I AM NOT DONE
 
-fn string_slice(arg: &str) {
-    println!("{}", arg);
+pub enum Command {
+    Uppercase,
+    Trim,
+    Append(usize),
 }
-fn string(arg: String) {
-    println!("{}", arg);
+
+mod my_module {
+    use super::Command;
+
+    // TODO: Complete the function signature!
+    pub fn transformer(input: ???) -> ??? {
+        // TODO: Complete the output declaration!
+        let mut output: ??? = vec![];
+        for (string, command) in input.iter() {
+            // TODO: Complete the function body. You can do it!
+        }
+        output
+    }
 }
 
-fn main() {
-    string_slice("blue");
-    string("red".to_string());
-    string(String::from("hi"));
-    string("rust is fun!".to_owned());
-    string("nice weather".into());
-    string(format!("Interpolation {}", "Station"));
-    string_slice(&String::from("abc")[0..1]);
-    string_slice("  hello there ".trim());
-    string("Happy Monday!".to_string().replace("Mon", "Tues"));
-    string("mY sHiFt KeY iS sTiCkY".to_lowercase());
+#[cfg(test)]
+mod tests {
+    // TODO: What do we need to import to have `transformer` in scope?
+    use ???;
+    use super::Command;
+
+    #[test]
+    fn it_works() {
+        let output = transformer(vec![
+            ("hello".into(), Command::Uppercase),
+            (" all roads lead to rome! ".into(), Command::Trim),
+            ("foo".into(), Command::Append(1)),
+            ("bar".into(), Command::Append(5)),
+        ]);
+        assert_eq!(output[0], "HELLO");
+        assert_eq!(output[1], "all roads lead to rome!");
+        assert_eq!(output[2], "foobar");
+        assert_eq!(output[3], "barbarbarbarbarbar");
+    }
 }
