@@ -1,6 +1,8 @@
 // structs2.rs
 // Address all the TODOs to make the tests pass!
+// Execute `rustlings hint structs2` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 #[derive(Debug)]
 struct Order {
@@ -33,15 +35,7 @@ mod tests {
     fn your_order() {
         let order_template = create_order_template();
         // TODO: Create your own order using the update syntax and template above!
-        let your_order = Order {
-            name: "Hacker in Rust".to_string(),
-            year: 2019,
-            made_by_phone: false,
-            made_by_mobile: false,
-            made_by_email: true,
-            item_number: 123,
-            count: 1
-        };
+        // let your_order =
         assert_eq!(your_order.name, "Hacker in Rust");
         assert_eq!(your_order.year, order_template.year);
         assert_eq!(your_order.made_by_phone, order_template.made_by_phone);
