@@ -2,7 +2,9 @@
 // Structs contain data, but can also have logic. In this exercise we have
 // defined the Package struct and we want to test some logic attached to it.
 // Make the code compile and the tests pass!
-// If you have issues execute `rustlings hint structs3`
+// Execute `rustlings hint structs3` or use the `hint` watch subcommand for a hint.
+
+// I AM NOT DONE
 
 #[derive(Debug)]
 struct Package {
@@ -14,24 +16,22 @@ struct Package {
 impl Package {
     fn new(sender_country: String, recipient_country: String, weight_in_grams: i32) -> Package {
         if weight_in_grams <= 0 {
-            // Raise error that weight can't be in negetive or zero
-            panic!("Weight can't be in negetive or zero");
+            panic!("Can not ship a weightless package.")
         } else {
-            return Package {
+            Package {
                 sender_country,
                 recipient_country,
                 weight_in_grams,
-            };
+            }
         }
     }
 
-    fn is_international(&self) -> bool {
-        self.sender_country != self.recipient_country
+    fn is_international(&self) -> ??? {
+        // Something goes here...
     }
 
-    fn get_fees(&self, cents_per_gram: i32) -> i32 {
+    fn get_fees(&self, cents_per_gram: i32) -> ??? {
         // Something goes here...
-        cents_per_gram * self.weight_in_grams
     }
 }
 
@@ -78,5 +78,6 @@ mod tests {
         let package = Package::new(sender_country, recipient_country, 1500);
 
         assert_eq!(package.get_fees(cents_per_gram), 4500);
+        assert_eq!(package.get_fees(cents_per_gram * 2), 9000);
     }
 }
