@@ -4,7 +4,8 @@
 // rustlings run tests1
 
 // This test has a problem with it -- make the test compile! Make the test
-// pass! Make the test fail! Execute `rustlings hint tests1` for hints :)
+// pass! Make the test fail!
+// Execute `rustlings hint tests1` or use the `hint` watch subcommand for a hint.
 
 // I AM NOT DONE
 
