@@ -1,6 +1,7 @@
 // tests2.rs
 // This test has a problem with it -- make the test compile! Make the test
-// pass! Make the test fail! Execute `rustlings hint tests2` for hints :)
+// pass! Make the test fail!
+// Execute `rustlings hint tests2` or use the `hint` watch subcommand for a hint.
 
 // I AM NOT DONE
 
