@@ -1,13 +1,10 @@
 // variables1.rs
-// Make me compile! Execute the command `rustlings hint variables1` if you want a hint :)
+// Make me compile!
+// Execute `rustlings hint variables1` or use the `hint` watch subcommand for a hint.
 
-// About this `I AM NOT DONE` thing:
-// We sometimes encourage you to keep trying things on a given exercise,
-// even after you already figured it out. If you got everything working and
-// feel ready for the next exercise, remove the `I AM NOT DONE` comment below.
-
+// I AM NOT DONE
 
 fn main() {
-    let x: i32 = 5;
+    x = 5;
     println!("x has the value {}", x);
 }
