@@ -1,10 +1,9 @@
 // variables3.rs
-// Make me compile! Execute the command `rustlings hint variables3` if you want a hint :)
+// Execute `rustlings hint variables3` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
 fn main() {
-    let mut x: i32 = 3;
-    println!("Number {}", x);
-    x = 5; // don't change this line
+    let x: i32;
     println!("Number {}", x);
 }
