@@ -1,9 +1,11 @@
 // variables5.rs
-// Make me compile! Execute the command `rustlings hint variables5` if you want a hint :)
+// Execute `rustlings hint variables5` or use the `hint` watch subcommand for a hint.
+
+// I AM NOT DONE
 
 fn main() {
     let number = "T-H-R-E-E"; // don't change this line
     println!("Spell a Number : {}", number);
-    let number = 3;
+    number = 3; // don't rename this variable
     println!("Number plus two is : {}", number + 2);
 }
