@@ -1,8 +1,9 @@
 // variables6.rs
-// Make me compile! Execute the command `rustlings hint variables6` if you want a hint :)
+// Execute `rustlings hint variables6` or use the `hint` watch subcommand for a hint.
 
+// I AM NOT DONE
 
-const NUMBER: i32 = 3;
+const NUMBER = 3;
 fn main() {
     println!("Number {}", NUMBER);
 }
